worked on the track lines and implemented the bar code scanner

+++ app/src/main/java/com/example/salubris/database/entities/Product.kt
@@ -9,7 +9,7 @@ data class Product(
     @PrimaryKey(autoGenerate = true)
     val uid: Int = 0,
     @ColumnInfo(name = "name") val name: String,
-    @ColumnInfo(name = "calories") val calories: Int = 0,
+    @ColumnInfo(name = "calories") val calories: Float = 0f,
     @ColumnInfo(name = "protein") val protein: Float = 0f,
     @ColumnInfo(name = "carbs") val carbs: Float = 0f,
     @ColumnInfo(name = "fats") val fats: Float = 0f
